etl_runner/requirements.txt
# Google Cloud Dependencies
google-cloud-bigquery==3.14.0
google-cloud-secret-manager==2.16.4

# Database Drivers
psycopg2-binary==2.9.9
pymysql==1.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# HTTP & API
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
